bug fixes and new data_version

+++ pychunkedgraph/ingest/ran_ingestion.py
@@ -216,7 +216,7 @@ def _create_atomic_chunk(args):
     for i_chunk, chunk_coord in enumerate(chunks):
         time_start = time.time()
 
-        create_atomic_chunk(im, chunk_coord, aff_dtype=aff_dtype, verbose=True)
+        create_atomic_chunk(im, chunk_coord, verbose=True)
 
         print(f"Layer 1 - {chunk_coord} - Job {i_block + 1} / {n_blocks} - "
               f"{i_chunk + 1} / {len(chunks)} -- %.3fs" %
